after review changes

+++ polymorphia-backend/src/main/java/com/agh/polymorphia_backend/service/project/ProjectGroupService.java
@@ -2,6 +2,7 @@ package com.agh.polymorphia_backend.service.project;
 
 import com.agh.polymorphia_backend.model.project.ProjectGroup;
 import com.agh.polymorphia_backend.model.user.student.Student;
+import com.agh.polymorphia_backend.repository.grade.GradeRepository;
 import com.agh.polymorphia_backend.repository.project.ProjectGroupRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -14,6 +15,7 @@ import java.util.List;
 @AllArgsConstructor
 public class ProjectGroupService {
     private final ProjectGroupRepository projectGroupRepository;
+    private final GradeRepository gradeRepository;
 
     public List<Student> getStudentsFromProjectGroup(ProjectGroup projectGroup) {
         return projectGroup.getAnimals().stream().map(animal -> animal.getStudentCourseGroupAssignment().getStudent())
@@ -26,10 +28,20 @@ public class ProjectGroupService {
     }
 
     public void save(ProjectGroup projectGroup) {
+        if (projectGroup.getId() != null && gradeRepository.hasGroupGrades(projectGroup)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nie można edytować ocenionej grupy projektowej.");
+        }
         projectGroupRepository.save(projectGroup);
     }
 
     public void delete(ProjectGroup projectGroup) {
+        if (gradeRepository.hasGroupGrades(projectGroup)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Nie można usunąć ocenionej grupy projektowej.");
+        }
         projectGroupRepository.delete(projectGroup);
+    }
+
+    public List<Long> getStudentIdsFromProjectGroup(ProjectGroup projectGroup) {
+        return projectGroupRepository.getStudentIdsByProjectGroup(projectGroup);
     }
 }
